// src/dora/common/functions.enum.ts
export enum EFunctions {
  FUNCTION_1_CLIENT_ONBOARDING = 'FUNCTION_1',
  FUNCTION_2_CLIENT_DATA_STORAGE_MANAGEMENT = 'FUNCTION_2',
  FUNCTION_3_TRANSFERRING_ASSETS = 'FUNCTION_3',
  FUNCTION_4_TRADING = 'FUNCTION_4',
  FUNCTION_5_PORT_MANAGEMENT = 'FUNCTION_5',
  FUNCTION_6_DASHBOARD = 'FUNCTION_6',
  FUNCTION_7_CLIENT_TX_MONITORING = 'FUNCTION_7',
  FUNCTION_8_RISK_MANAGEMENTS = 'FUNCTION_8',
  FUNCTION_9_HR_TALENT_MANAGEMENT = 'FUNCTION_9',
  FUNCTION_10_RECORD_KEEPING_DOCS = 'FUNCTION_10',
  FUNCTION_11_MARKETING_COMMUNICATION = 'FUNCTION_11',
  FUNCTION_12_EMAILING = 'FUNCTION_12',
  FUNCTION_13_IT_SYSTEM_MNGMNT = 'FUNCTION_13',
  FUNCTION_14_INVOICING = 'FUNCTION_14',
}